Fallback to hwloc fetch if package not found
In case of dynamic hwloc library linking, fallback to fetching hwloc
from source instead of failing UMF build.

Add a workflow for testing the fallback to build a static hwloc library
in case it's not found in the system.

Always build a Release version of hwloc on Windows.

Unify correct paths to hwloc.

Refactor build steps dependencies.